Mapping Cassandra Rows to Entity Object in application

+++ src/main/java/com/app/model/AbstractEntity.java
@@ -9,14 +9,14 @@ import java.io.Serializable;
  */
 public abstract class AbstractEntity implements Serializable {
 
-    protected int id;
+    protected long id;
 
     @PrimaryKey
-    public BigInt getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
