+++ src/main/java/com/app/repository/UserInfoDao.java
@@ -5,6 +5,6 @@ import com.app.model.UserInfo;
 /**
  * Created by cassandra on 6/9/14.
  */
-public interface UserInfoDao extends AbstractCassandraDao<UserInfo, Integer> {
+public interface UserInfoDao extends AbstractCassandraDao<UserInfo, Long> {
 
 }
